gestion des opérations

+++ GR2_Projet/Model/Operation.cs
@@ -86,8 +86,8 @@ namespace GR2_Projet.Model
         /// <summary>
         /// Propriété montant.
         /// </summary>
-        private int mAmount;
-        public int Amount
+        private float mAmount;
+        public float Amount
         {
             get { return mAmount; }
             set { mAmount = value; }
@@ -96,7 +96,7 @@ namespace GR2_Projet.Model
         /// <summary>
         /// Constructeur.
         /// </summary>
-        public Operation(string Name, operationType OperationType, DateTime Date, paymentType PaymentType, int Amount)
+        public Operation(string Name, operationType OperationType, DateTime Date, paymentType PaymentType, float Amount)
         {
             this.Id = Guid.NewGuid().ToString();
             this.Name = Name;
